requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
requests
langchain
pymupdf